+++ app/src/main/java/org/ole/planet/myplanet/model/JoinedMemberData.kt
@@ -1,0 +1,10 @@
+package org.ole.planet.myplanet.model
+
+data class JoinedMemberData(
+    val user: UserEntity,
+    val visitCount: Long,
+    val lastVisitDate: Long?,
+    val offlineVisits: String,
+    val profileLastVisit: String,
+    var isLeader: Boolean
+)
